feat: 검색 결과·상세 페이지 동적 메타데이터 추가
